fix: rename nats -> stan (as it is what it actually implements)
BPS URL scheme also changed for it nats:// -> stan://
